refactor: home component_select_point with the other point gadgets
Its documentation links [`TorsionFreeWitnessPoint`], which is in scope
in `point` but not in `select`; the mux is a point operation, so it
belongs there anyway.
